Warns when a file can't be saved
Previously, Parchment would only warn when attempting to save a file
that had been modified by an external program—i.e. when its timestamp
was different than when it was last accessed by Parchment. Now, if a
file fails to write, Parchment attempts to discern the reason why and
will create a popup warning the user of the reason, offering either to
save the file to another path, to try again (after the user has
presumably rectified the error, if possible), or to abort for now.