EXPERIMENT: Indent when soft-wrapping
This commit adds code which enables Parchment to indent lines based on a
few heuristics when soft-wrapping, but the functionality is disabled due
to performance and correctness concerns.

The soft-wrapping includes heuristics for comments in C/C++ and Lua, as
well as for prefixes common in Markdown, allowing quotes and bullet
points to be displayed as a block.

The aforementioned performance issues are that it's very time consuming
to mark hundreds of lines at once, such as when loading or when
replacing all of a file's contents. Changing the logic to run only when
making small edits causes some cases where wrap-indentation might not
correctly re-run.

Code is being kept because the editing benefits of this feature, would
it to work correctly, would be astounding; It's worth trying again with
what has already been accomplished.